+++ hc/front/forms.py
@@ -122,6 +122,15 @@ class AddWebhookForm(forms.Form):
         max_length=1000, required=False, validators=[WebhookValidator()]
     )
 
+    def clean(self):
+        super().clean()
+
+        url_down = self.cleaned_data.get("url_down")
+        url_up = self.cleaned_data.get("url_up")
+
+        if not url_down and not url_up:
+            self.add_error("url_down", "Enter a valid URL.")
+
     def get_value(self):
         return json.dumps(dict(self.cleaned_data), sort_keys=True)