+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - Fix List-Unsubscribe email header value: add angle brackets
 - Unsubscribe links serve a form, and require HTTP POST to actually unsubscribe
 - For webhook integration, validate each header line separately
+- Fix "Send Test Notification" for webhooks that only fire on checks going up
 
 
 ## v1.11.0 - 2019-11-22